add: new rest architecture
