+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -3,7 +3,7 @@ package ru.yandex.practicum.filmorate.controller;
 import jakarta.validation.Valid;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.service.UserService;
+import ru.yandex.practicum.filmorate.storage.user.InMemoryUserStorage;
 
 import java.util.Collection;
 
@@ -11,24 +11,24 @@ import java.util.Collection;
 @RequestMapping("/users")
 public class UserController {
 
-    private final UserService userService;
+    private final InMemoryUserStorage inMemoryUserStorage;
 
-    public UserController(UserService userService) {
-        this.userService = userService;
+    public UserController(InMemoryUserStorage inMemoryUserStorage) {
+        this.inMemoryUserStorage = inMemoryUserStorage;
     }
 
     @GetMapping
     public Collection<User> getAllUsers() {
-        return userService.getAllUsers();
+        return inMemoryUserStorage.getAllUsers();
     }
 
     @PostMapping
     public User create(@Valid @RequestBody User user) {
-        return userService.create(user);
+        return inMemoryUserStorage.create(user);
     }
 
     @PutMapping
     public User update(@Valid @RequestBody User newUser) {
-        return userService.update(newUser);
+        return inMemoryUserStorage.update(newUser);
     }
 }
