+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,69 +1,7 @@
 package ru.yandex.practicum.filmorate.service;
 
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.exceptions.ValidationException;
-import ru.yandex.practicum.filmorate.model.Film;
 
-import java.time.LocalDate;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
-
-@Slf4j
 @Service
 public class FilmService {
-
-    private final Map<Integer, Film> films = new HashMap<>();
-
-    public Collection<Film> getAllFilms() {
-        return films.values();
-    }
-
-    public Film create(Film film) {
-        validateFilm(film);
-        film.setId(getNextId());
-        films.put(film.getId(), film);
-        log.info("Добавлен фильм {}", film);
-        return film;
-    }
-
-    public Film update(Film newFilm) {
-        if (newFilm.getId() == null) {
-            throw new ValidationException("Id должен быть указан");
-        }
-        if (!films.containsKey(newFilm.getId())) {
-            throw new ValidationException("Фильм с id = " + newFilm.getId() + " не найден");
-        }
-        validateFilm(newFilm);
-        films.put(newFilm.getId(), newFilm);
-        log.info("Изменен фильм {}", newFilm);
-        return newFilm;
-
-    }
-
-    void validateFilm(Film film) {
-        boolean isNoValidRelease = film.getReleaseDate().isBefore(LocalDate.of(1895, 12, 28));
-        if (film.getName() == null) {
-            throw new ValidationException("Название не может быть пустым");
-        }
-        if (film.getDescription().length() > 200) {
-            throw new ValidationException("Описание не может быть больше 200 символов");
-        }
-        if (isNoValidRelease) {
-            throw new ValidationException("Некорректная дата релиза фильма");
-        }
-        if (film.getDuration() < 0) {
-            throw new ValidationException("Продолжительность фильма должна быть положительным числом");
-        }
-    }
-
-    private int getNextId() {
-        int currentMaxId = films.keySet()
-                .stream()
-                .mapToInt(id -> id)
-                .max()
-                .orElse(0);
-        return ++currentMaxId;
-    }
 }
