+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/InMemoryUserStorage.java
@@ -1,0 +1,76 @@
+package ru.yandex.practicum.filmorate.storage.user;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.filmorate.exceptions.ValidationException;
+import ru.yandex.practicum.filmorate.model.User;
+
+import java.time.LocalDate;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.Map;
+
+@Slf4j
+@Component
+public class InMemoryUserStorage implements UserStorage {
+
+    private final Map<Integer, User> users = new HashMap<>();
+
+    @Override
+    public Collection<User> getAllUsers() {
+        return users.values();
+    }
+
+    @Override
+    public User create(User user) {
+        validateUser(user);
+        user.setId(getNextId());
+        if (user.getName() == null || user.getName().isBlank()) {
+            user.setName(user.getLogin());
+        }
+        users.put(user.getId(), user);
+        log.info("Добавлен пользователь {}", user.getName());
+        return user;
+    }
+
+    @Override
+    public User update(User newUser) {
+        if (newUser.getId() == null) {
+            throw new ValidationException("Id должен быть указан");
+        }
+        if (!users.containsKey(newUser.getId())) {
+            throw new ValidationException("Фильм с id = " + newUser.getId() + " не найден");
+        }
+        validateUser(newUser);
+        if (newUser.getName() == null || newUser.getName().isBlank()) {
+            newUser.setName(newUser.getLogin());
+        } else {
+            newUser.setName(newUser.getName());
+        }
+        users.put(newUser.getId(), newUser);
+        log.info("Изменен пользователь {}", newUser.getName());
+        return newUser;
+    }
+
+    void validateUser(User user) {
+        boolean isNoValidBirthday = user.getBirthday().isAfter(LocalDate.now());
+        if (user.getEmail().isEmpty() || !user.getEmail().contains("@")) {
+            throw new ValidationException("Электронная почта не может быть пустой и должна содержать символ '@'");
+        }
+        if (user.getLogin().isEmpty() || user.getLogin().contains(" ")) {
+            throw new ValidationException("Логин не может быть пустым и содержать пробелы");
+        }
+        if (isNoValidBirthday) {
+            throw new ValidationException("Некорректная дата рождения");
+        }
+    }
+
+    private int getNextId() {
+        int currentMaxId = users.keySet()
+                .stream()
+                .mapToInt(id -> id)
+                .max()
+                .orElse(0);
+        return ++currentMaxId;
+    }
+}
